requirements.txt
langchain>=0.3.0
langchain-openai>=0.1.0
langchain-mcp-adapters>=0.1.0
fastmcp>=2.10.0
akshare>=1.11.0
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
requests>=2.25.0